move scss files into header
