+++ src/components/Header/account-dashboard.scss
@@ -1,7 +1,3 @@
-@use "../_constants.scss";
-@use "../_variables.scss";
-@use "../_mixins.scss";
-
 @media (max-width: 670px) {
     .account-dashboard {
         grid-template-columns: repeat(1, 1fr);
@@ -13,29 +9,23 @@
     .account-dashboard {
         grid-template-columns: repeat(2, 1fr);
         height: 23rem;
-
     }
 }
 
 .account-dashboard {
 
-
     .tab-container {
         width: 70%;
         margin: auto;
-
     }
 
     .tab-selector {
         margin: -0.5rem auto 0;
         max-width: 120rem;
-        //width: 80%;
-
     }
 
     .tabbed-content {
         padding-top: 1rem;
         height: 22.5rem;
-
     }
 }
