+++ src/components/Header/header.scss
@@ -1,7 +1,5 @@
-@use "../_constants.scss";
-@use "../_variables.scss";
-@use "../_mixins.scss";
-
+@use "../../styles/mixins";
+@use "../../styles/constants";
 
 .header {
     button:hover {
