+++ src/components/Header/main-summary.scss
@@ -1,0 +1,3 @@
+.main-summary-div {
+    border: 2px solid orange;
+}
