+++ src/components/Header/secondary-summary.scss
@@ -1,7 +1,3 @@
-@use "../_constants.scss";
-@use "../_variables.scss";
-@use "../_mixins.scss";
-
 .secondary-summary-div {
     .overflow-table {
         overflow-x: scroll;
