--- src/styles/componentStyles/main-summary.scss
@@ -1,7 +1,0 @@
-@use "../_constants.scss";
-@use "../_variables.scss";
-@use "../_mixins.scss";
-
-.main-summary-div {
-    border: 2px solid orange;
-}
